+++ R/wrapper-feasible-and-infeasible-region-search.R
@@ -64,9 +64,9 @@
 #' 
 #' @references
 #' 
-#' Xiaofan Wu et al. “Feasible and infeasible region search for # nolint
-#' the maximally diverse grouping problem”. In: Computers & Operations Research
-#' 179 (2025). DOI: https://doi.org/10.1016/j.cor.2025.107030. 
+#' Wu, X., Feng, J., Yang, J., & Zhang, Y. (2025). Feasible and infeasible region 
+#' search for the maximally diverse grouping problem. Computers & Operations Research, 179,
+#' 107030. https://doi.org/10.1016/j.cor.2025.107030
 #'
 feasible_and_infeasible_region_search_anticlustering <- function(
     x, K, N, objective = "diversity", number_iterations=50, clusters=NULL, upper_bound=NULL,